Fix Docker compatability
