+++ Cargo.toml
@@ -18,4 +18,4 @@ mime = "0.3"
 num = "0.2"
 custom_error = "1.7"
 libwebp-sys = "0.3"
-addr = "0.2"
+dns-lookup = "1.0"
