+++ Dockerfile
@@ -1,5 +1,4 @@
-FROM ekidd/rust-musl-builder AS builder
-RUN sudo apt-get update -y
+FROM ekidd/rust-musl-builder:nightly-2020-04-10 AS builder
 RUN USER=rust:rust cargo init . --name mediaproxy
 COPY --chown=rust:rust Cargo.* ./
 RUN cargo build --release
