+++ nginx.conf
@@ -102,7 +102,8 @@ http {
       expires 1d;
       proxy_cache cache_zone;
       proxy_cache_valid 200 302 301 1d;
-      # proxy_cache_key $scheme://$host$request_uri$args;
+      proxy_cache_key "$request_uri|$request_body";
+      proxy_cache_methods POST;
       proxy_pass http://mediaproxy;
       proxy_set_header Host $http_host;
       proxy_set_header X-Real-IP $remote_addr;
