+++ src/fetching/mod.rs
@@ -1,0 +1,4 @@
+pub mod validation;
+mod fetching;
+
+pub use self::fetching::*;
